requirements.txt
numpy
h5py
pillow
tensorflow>=1.5.0
keras>=2.1.3
